separate app and server into different files
